update: setup.sh

+++ setup.sh
@@ -1,21 +1,18 @@
-#!/bin/bash
+#! /bin/bash
+
+set -eu
+
+script_home="$(cd $(dirname $0); pwd)"
+cd ${script_home}
 
 for f in .??*
 do
-	[[ "$f" == ".git" ]] && continue
-	[[ "$f" == ".DS_Store" ]] && continue
+  [[ "$f" == ".git" ]] && continue
+  [[ "$f" == ".DS_Store" ]] && continue
 
-	echo "$f"
-	unlink $HOME/$f
-	ln -s $(cd $(dirname $0); pwd)/$f $HOME/$f
+  if [ -e "$HOME/$f" ]; then
+    unlink $HOME/$f
+  fi
+  ln -s $(cd $(dirname $0); pwd)/$f $HOME/$f
+  echo "created symbolic link: $f"
 done
-
-chsh -s `which zsh`
-git clone --recursive https://github.com/sorin-ionescu/prezto.git "${ZDOTDIR:-$HOME}/.zprezto"
-
-if [ $(uname) = 'Linux' ]; then
-	curl -fsSL https://starship.rs/install.sh | bash
-else
-	brew install starship
-fi
-
